+++ addOns/authhelper/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Exception in authentication diagnostics.
 
+### Changed
+- Maintenance changes.
+
 ## [0.34.0] - 2025-12-15
 ### Changed
 - Update minimum ZAP version to 2.17.0.